Architecture fix
